Elimino layouts que ya no se usaban, refactor de algunos mensajes y titulos. Mejora de layouts en general
