+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:theme="@style/AppTheme">
         <activity android:name=".ui.userVideos.UserVideosActivity"></activity>
         <activity
-            android:name=".ui.myUserProfile.EditProfileActivity"
+            android:name=".ui.userProfile.EditProfileActivity"
             android:label="@string/edit_profile_activity" />
         <activity
             android:name=".ui.recovery.RecoveryActivity"
