+++ app/src/main/java/com/example/chotuve_android_client/ui/notifications/NotificationsFragment.kt
@@ -26,7 +26,7 @@ class NotificationsFragment : Fragment() {
                 ViewModelProviders.of(this).get(NotificationsViewModel::class.java)
         val root = inflater.inflate(R.layout.fragment_notifications, container, false)
         val textView: TextView = root.findViewById(R.id.text_notifications)
-        notificationsViewModel.text.observe(viewLifecycleOwner, Observer {
+        notificationsViewModel.title.observe(viewLifecycleOwner, Observer {
             textView.text = it
         })
