+++ app/src/main/java/com/example/chotuve_android_client/ui/userProfile/MyUserProfileFragment.kt
@@ -1,9 +1,7 @@
-package com.example.chotuve_android_client.ui.myUserProfile
+package com.example.chotuve_android_client.ui.userProfile
 
 import android.content.Context
-import android.content.Context.MODE_PRIVATE
 import android.content.Intent
-import android.content.SharedPreferences
 import android.os.Bundle
 import android.util.Log
 import android.view.*
@@ -12,11 +10,9 @@ import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
-import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.viewpager.widget.ViewPager
 import com.example.chotuve_android_client.MainActivity
 import com.example.chotuve_android_client.R
-import com.example.chotuve_android_client.tools.adapters.FriendsAdapter
 import com.example.chotuve_android_client.tools.adapters.PageAdapter
 import com.example.chotuve_android_client.ui.login.CHOTUVE_SHARED_PREFS
 import com.example.chotuve_android_client.ui.login.LoginActivity
