+++ app/src/main/res/layout/fragment_my_user_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.myUserProfile.MyUserProfileFragment">
+    tools:context=".ui.userProfile.MyUserProfileFragment">
 
 
     <de.hdodenhof.circleimageview.CircleImageView
