+++ app/src/main/res/layout/fragment_videos_view_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.myUserProfile.FriendsViewPagerFragment">
+    tools:context=".ui.userProfile.FriendsViewPagerFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerview_profile_videos"
