+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -31,7 +31,7 @@
 
     <fragment
         android:id="@+id/navigation_user_profile"
-        android:name="com.example.chotuve_android_client.ui.myUserProfile.MyUserProfileFragment"
+        android:name="com.example.chotuve_android_client.ui.userProfile.MyUserProfileFragment"
         android:label="@string/title_user"
         tools:layout="@layout/fragment_my_user_profile" />
